+++ crates/rils_execution/src/value/display.rs
@@ -42,6 +42,8 @@ impl fmt::Display for Value {
             Self::BinaryHeap(_) => write!(f, "<BinaryHeap>"),
             Self::SequenceIterator(_) => write!(f, "<sequence iterator>"),
             Self::BorrowedSequenceIterator(_) => write!(f, "<borrowed sequence iterator>"),
+            Self::BorrowedMapIterator(_) => write!(f, "<borrowed map iterator>"),
+            Self::BorrowedSetIterator(_) => write!(f, "<borrowed set iterator>"),
             Self::BytecodeIterator(_) => write!(f, "<bytecode iterator>"),
             Self::Reference(reference) => match reference.read() {
                 Ok(value) => write!(f, "{value}"),