+++ crates/rils_execution/src/value/hash.rs
@@ -84,25 +84,88 @@ impl HashKey {
 }
 
 pub struct HashMapValue {
+    pub borrowed: std::cell::Cell<usize>,
     pub entries: RefCell<HashMap<HashKey, FieldSlot>>,
     pub key_type: RefCell<Type>,
     pub value_type: RefCell<Type>,
 }
 
 pub struct BTreeMapValue {
+    pub borrowed: std::cell::Cell<usize>,
     pub entries: RefCell<BTreeMap<HashKey, FieldSlot>>,
     pub key_type: RefCell<Type>,
     pub value_type: RefCell<Type>,
 }
 
 pub struct BTreeSetValue {
+    pub borrowed: std::cell::Cell<usize>,
     pub entries: RefCell<BTreeSet<HashKey>>,
     pub element_type: RefCell<Type>,
 }
 
 pub struct HashSetValue {
+    pub borrowed: std::cell::Cell<usize>,
     pub entries: RefCell<HashSet<HashKey>>,
     pub element_type: RefCell<Type>,
+}
+
+#[derive(Clone)]
+pub enum MapCollection {
+    Hash(Rc<HashMapValue>),
+    BTree(Rc<BTreeMapValue>),
+}
+
+impl MapCollection {
+    pub fn borrowed(&self) -> &std::cell::Cell<usize> {
+        match self {
+            Self::Hash(map) => &map.borrowed,
+            Self::BTree(map) => &map.borrowed,
+        }
+    }
+
+    pub fn contains_key(&self, key: &HashKey) -> bool {
+        match self {
+            Self::Hash(map) => map.entries.borrow().contains_key(key),
+            Self::BTree(map) => map.entries.borrow().contains_key(key),
+        }
+    }
+
+    pub fn value(&self, key: &HashKey) -> Option<Value> {
+        match self {
+            Self::Hash(map) => map
+                .entries
+                .borrow()
+                .get(key)
+                .and_then(|slot| slot.value.clone()),
+            Self::BTree(map) => map
+                .entries
+                .borrow()
+                .get(key)
+                .and_then(|slot| slot.value.clone()),
+        }
+    }
+}
+
+#[derive(Clone)]
+pub enum SetCollection {
+    Hash(Rc<HashSetValue>),
+    BTree(Rc<BTreeSetValue>),
+}
+
+impl SetCollection {
+    pub fn borrowed(&self) -> &std::cell::Cell<usize> {
+        match self {
+            Self::Hash(set) => &set.borrowed,
+            Self::BTree(set) => &set.borrowed,
+        }
+    }
+
+    pub fn contains(&self, key: &HashKey) -> bool {
+        match self {
+            Self::Hash(set) => set.entries.borrow().contains(key),
+            Self::BTree(set) => set.entries.borrow().contains(key),
+        }
+    }
 }
 
 pub(super) fn clone_hash_map(map: &HashMapValue) -> Result<HashMapValue, String> {
@@ -126,6 +189,7 @@ pub(super) fn clone_hash_map(map: &HashMapValue) -> Result<HashMapValue, String>
         })
         .collect::<Result<HashMap<_, _>, String>>()?;
     Ok(HashMapValue {
+        borrowed: std::cell::Cell::new(0),
         entries: RefCell::new(entries),
         key_type: RefCell::new(map.key_type.borrow().clone()),
         value_type: RefCell::new(map.value_type.borrow().clone()),
@@ -153,6 +217,7 @@ pub(super) fn clone_btree_map(map: &BTreeMapValue) -> Result<BTreeMapValue, Stri
         })
         .collect::<Result<BTreeMap<_, _>, String>>()?;
     Ok(BTreeMapValue {
+        borrowed: std::cell::Cell::new(0),
         entries: RefCell::new(entries),
         key_type: RefCell::new(map.key_type.borrow().clone()),
         value_type: RefCell::new(map.value_type.borrow().clone()),